pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "linear-manager"
version = "0.1.0"
description = "CLI tool to sync YAML-defined tickets to Linear"
readme = "README.md"
requires-python = ">=3.12"
authors = [{name = "LinearManager Maintainers"}]
license = {text = "MIT"}
dependencies = [
    "httpx>=0.27",
    "PyYAML>=6.0",
    "colorama>=0.4.6",
]
keywords = ["linear", "cli", "automation"]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.1",
]

[project.scripts]
manager = "linear_manager.cli:main"

[tool.uv]
package = true
